change weight from int to float on sequelize migration

+++ db/migrations/20220517191541-add-weight-to-tx.js
@@ -3,7 +3,7 @@
 module.exports = {
   up: async (queryInterface, Sequelize) => {
     await queryInterface.addColumn('transaction', 'weight', {
-      type: Sequelize.INTEGER.UNSIGNED,
+      type: Sequelize.FLOAT.UNSIGNED,
       allowNull: true,
       defaultValue: null,
     });
